refactored odds_alert.py into formatter and pre start check methods/functions/utils
